requirements.txt
# Core
python-dotenv==1.0.1
httpx==0.27.2

# Data & Finance
pandas==2.2.3

# AI
# DSPy (uses LiteLLM internally for provider adapters)
dspy-ai==2.5.29
backoff==2.2.1
joblib==1.4.2
datasets==2.20.0
ujson==5.10.0
magicattr==0.1.6

# Visualization
plotly==5.24.1
streamlit==1.38.0
